requirements.txt
Flask==3.1.*
Flask-Login==0.6.*
Flask-Bcrypt==1.0.*
Flask-SQLAlchemy==3.1.*
Flask-WTF==1.2.*
Flask-Compress==1.*
gunicorn==23.*
python-dotenv==1.*
markdown==3.*
resend
boto3
